enable switching between reading FASTA into memory and indexed unbuffered random access
